requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
Deprecated==1.2.13
idna==3.2
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
QtAwesome==1.0.3
QtPy==1.11.2
requests==2.26.0
six==1.16.0
urllib3==1.26.6
wrapt==1.12.1
